+++ lib/active_interactor/context/attributes.rb
@@ -27,16 +27,17 @@ module ActiveInteractor
         #
         # @return [Array<Symbol>] the defined attributes
         def attributes(*attributes)
-          return __attributes if attributes.empty?
+          attributes.compact.uniq.each { |attr| attribute(attr) }
 
-          @__attributes = __attributes.concat(attributes.map(&:to_sym)).compact.uniq.sort
+          attribute_names.sort.collect(&:to_sym)
         end
 
         private
 
-        def __attributes
-          @__attributes ||= []
+        def attribute?(attr_name)
+          attribute_types.key?(attr_name.to_s)
         end
+        alias has_attribute? attribute?
       end
 
       # Initialize a new instance of {Base}
@@ -48,6 +49,8 @@ module ActiveInteractor
         copy_flags!(context)
         copy_called!(context)
         super
+
+        merge_attribute_values(context)
       end
 
       # Get values defined on the instance of {Base context} whose keys are defined on the {Base context} class'
@@ -69,10 +72,19 @@ module ActiveInteractor
       #
       # @return [Hash{Symbol => *}] the defined attributes and values
       def attributes
-        self.class.attributes.each_with_object({}) do |attribute, hash|
-          hash[attribute] = self[attribute] if self[attribute]
-        end
+        super.symbolize_keys
       end
+
+      # Check if the {Base context} instance has an attribute
+      #
+      # @since unreleased
+      #
+      # @param attr_name [Symbol, String] the name of the attribute to check
+      # @return [Boolean] whether or not the {Base context} instance has the attribute
+      def attribute?(attr_name)
+        @attributes.key?(attr_name.to_s)
+      end
+      alias has_attribute? attribute?
 
       # Merge an instance of {Base context} into the calling {Base context} instance
       #
@@ -96,6 +108,13 @@ module ActiveInteractor
       end
 
       private
+
+      def merge_attribute_values(context)
+        return unless context
+
+        values = context.respond_to?(:attributes) ? context.attributes : context
+        values.each { |key, value| public_send("#{key}=", value) }
+      end
 
       def merge_errors!(errors)
         if errors.is_a? String