+++ CHANGELOG.md
@@ -7,6 +7,29 @@ and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
 
+### Added
+
+- `ActiveInteractor::Context::Attributes#attribute?`
+- `ActiveInteractor::Context::Attributes#has_attribute?`
+- `ActiveInteractor::Context::Base#attribute_missing`
+- `ActiveInteractor::Context::Base#attribute_names`
+- `ActiveInteractor::Context::Base#method_missing`
+- `ActiveInteractor::Context::Base#respond_to?`
+- `ActiveInteractor::Context::Base#respond_to_without_attributes?`
+- `ActiveInteractor::Context::Base.attribute`
+- `ActiveInteractor::Context::Base.attribute_missing`
+- `ActiveInteractor::Context::Base.attribute_names`
+- `ActiveInteractor::Context::Base.method_missing`
+- `ActiveInteractor::Context::Base.respond_to?`
+- `ActiveInteractor::Context::Base.respond_to_without_attributes?`
+- `ActiveInteractor::Interactor::Context#context_attribute_missing`
+- `ActiveInteractor::Interactor::Context#context_attribute_names`
+- `ActiveInteractor::Interactor::Context#context_respond_to_without_attributes?`
+- `ActiveInteractor::Interactor::Context::ClassMethods#context_attribute`
+- `ActiveInteractor::Interactor::Context::ClassMethods#context_attribute_missing`
+- `ActiveInteractor::Interactor::Context::ClassMethods#context_attribute_names`
+- `ActiveInteractor::Interactor::Context::ClassMethods#context_respond_to_without_attributes?`
+
 ## [v1.0.0] - 2020-01-26
 
 ### Added